Fixed Sunrise and Sunset times.
Conversion of sunrise and sunset timestamps to a legible date uses the
local timezone, and therefore will show local time for them. Using
Google Timezones API, we can get the timezone of the place and modify
the timestamps with the correct offset.

The offset is based on the GMT, so the sunrise and sunset will still be
wrong if we're not at that timezone. So we need to get the local
timezone and apply the correction, besides the offset we got from the
API.