Added warning when adding markers twice
If the user tries to put a marker on a city that already has one, an
alert dialog will pop up. It will warn the user about an existing marker
on that city, and will give him the possibility to go to that marker.
This works both for the search function, and for long click operations
on the map.