Added unit converters, reduced OWM queries.
- When changing units in the settings, the app will modify each weather
object using math operations, not asking the API with the new
parameters.
- When clicking on a marker in the map, the app doesn't update the
weather automatically. There will be an update button later.

Because of this last modification, an unsaved marker weather would be
lost if you clicked on a saved marker. There's a new variable that has
the unsaved marker weather, JIC the user clicks on it again.